requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
mcp>=1.0.0
langsmith>=0.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
